Fix some radio stations not playing with the relaying
The problem was how we handled the HTTP headers in case the original stream
URL redirected the connection to another URL. And specifically, if the
redirection response returned the Content-Length header with value 0 and the
final URL didn't have this header (as is typical for live streams). In that
case, the 0-length was erroneously returned to the client as the length of
the stream.

The problem originated from how the "associative" mode of the PHP built-in
function get_headers works. It makes it basically impossible to know for
certain, which headers come from the final URL and which ones from the
intermediate redirection response(s). To work around this, we now use the
function in non-associative mode and parse the result to a dictionary with
our own logic.

refs https://github.com/owncloud/music/issues/1194